codon_harmonizer/requirements.txt
bcbio-gff==0.7.0
beautifulsoup4==4.12.2
Bio==1.5.9
ete3==3.1.2
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
requests==2.28.2
seaborn==0.12.2
lxml==4.9.2
beartype-0.15.0
